Did testing on Playlist and Room (OLD MODEL)
Created a new testing document for anything related to Playlist and Room as the first testing file was very long and crowded. 

This was done with the model used for deliverable 1 and the view has not been changed to incorporate these new methods as of yet. (Have not added a Create New Playlist, Add Song to Playlist or Add Room view).